requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-socketio==5.3.6
flask-jwt-extended==4.6.0
pymysql==1.1.1
python-dotenv==1.0.1
apscheduler==3.10.4
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
openpyxl==3.1.2
eventlet==0.36.1
feedparser==6.0.10
vaderSentiment==3.3.2
